refactor(assistant): split tool registry files (refs #246)
tools.ts had grown to 292 LOC holding safeExecute, the navigate
allowlist, all 9 read-only tools, and the registry barrel. Adding the
destructive tools from #246 would push it past the ~400 LOC cap
(AGENTS.md rule 12), per Task 3 review.

- tool-helpers.ts: safeExecute + NAVIGATE_ALLOWLIST, shared by every tool file
- tools-readonly.ts: the 9 read-only ToolDefinitions, exporting readOnlyTools
- tools.ts: barrel only, assembling TOOLS and getToolByName

Also hardens list_events' limit clamp: a non-numeric or negative
input.limit now falls back to ASSISTANT.maxListEventsLimit via
Number.isFinite instead of producing a NaN/negative EventFilters.limit
(Task 3 review minor).
